SQL corrections (#8)
* Adding views

Adding a more descriptive view of the slowly changing dimensions table of Dicounts of the bought books. And Invoice and Invoice Details viewm that report on the books bought as a subtotal, their lowered price, and the total amount pewr transaction.

* Change from 'Create <Name_of_table>' to 'Table  <Name_of_table>'

Changes the names of the create table sql files from 'Create <Name_of_table>.sql' to 'Table  <Name_of_table>.sql'

* More views

Adding a view of the sales in the last two months (by user and in general). As well as an extra view of the avaliable data of the translated books xoalesced with the data in the original language (froom the Books table).

* Create insert.sh

Adding a insert bash script to create insert into sql statements from the .csv files with the data.

* Small fixes in SQL queries

Editing the queries by correcting grammar isues, adding SQL statements and their order, and adding the CompleteDiscountView to see future discounts (arbitrarily to 10 years).